requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pdfplumber>=0.11.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
